Partial merge r178:179 of branches/sunflower--dwa-configurable: * Removed api/types.hpp, using size_t throughout the library now. * Added HullIterator, marked Hull::GetLine() as deprecated. * Factored-out LineRayIntersect() (from util/Ray) and   LineCircleIntersect() (from util/Line) into util/numeric for easier   use elsewhere without the need for temporary instances. * Tweaked depend / redepend target (avoid frequent depend rebuilts). * api/RobotModel: added m_safety_hull, GetSafetyHull(), internally   renamed securityDistance ==> safetyDistance * dwa/DistanceObjective: comment tweaks (typo, todo)
git-svn-id: https://libsunflower.svn.sourceforge.net/svnroot/libsunflower/trunk/sunflower@190 86aca762-15d7-0310-9726-9bd8562223d8

+++ sfl/util/Polygon.hpp
@@ -82,7 +82,7 @@ namespace sfl {
     */
     double CalculateRadius() const;
     
-    unsigned int GetNPoints() const { return m_point.size(); }
+    size_t GetNPoints() const { return m_point.size(); }
     
     /**
        Convenience method for legacy code.
@@ -90,10 +90,10 @@ namespace sfl {
        \return A copy of the line between the corners (index) and
        (index + 1). The last point is connected with the first one.
     */
-    boost::shared_ptr<Line> GetLine(int index) const;
+    boost::shared_ptr<Line> GetLine(size_t index) const;
     
     /** \return Pointer to the point at given index, or 0 if invalid index. */
-    const Point * GetPoint(unsigned int index) const;
+    const Point * GetPoint(size_t index) const;
     
     /**
        Writes the corners in human readable format on the provided ostream.
